requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
Pillow==10.4.0
python-decouple==3.8
whitenoise==6.7.0
gunicorn==22.0.0
dj-database-url==2.2.0
psycopg2-binary==2.9.9
django-filter==24.3
